requirements.txt
python-dotenv==1.0.0
langchain==0.0.308
azure-ai-formrecognizer==3.2.0
azure-search-documents==11.4.0b8
azure-identity==1.13.0
fastapi
uvicorn